+++ lib/helpers/inventory_helpers.py
@@ -1,26 +1,26 @@
 import os
-from typing import Dict, List, Tuple
+from typing import List, Tuple
 
-from .classes import Proxy
-from .os_helpers import duplicate_file
+from lib.classes import InventoryCard
+from lib.helpers import os_helpers
 
 
 def create_unique_proxies(
-    proxies: List[Proxy],
+    proxies: List[InventoryCard],
     proxy_folder: str,
     destination_folder: str,
     duplicate: bool = True,
-) -> Tuple[List[Proxy], List[Proxy]]:
+) -> Tuple[List[InventoryCard], List[InventoryCard]]:
     """Traverse folder of proxies and create unique proxy images.
 
     Args:
-        proxies (List[Proxy]): List of proxies to create files for
+        proxies (List[InventoryCard]): List of proxies to create files for
         proxy_folder (str): Folder of source proxies
         destination_folder (str): Folder to write unique files for
         duplicate (bool): Flag to duplicate file based on proxy count. Defaults to True.
 
     Returns:
-        Tuple[List[Proxy], List[Proxy]]: List of missing files and duplicate files respectively
+        Tuple[List[InventoryCard], List[InventoryCard]]: List of missing files and duplicate files respectively
     """
     errors = { "missing": [], "duplicates": [] }
     for proxy in proxies:
@@ -36,6 +36,7 @@ def create_unique_proxies(
             continue
 
         proxy_path = f"{proxy_folder}/{proxy.type}/{file_names[0]}"
-        duplicate_file(proxy_path, destination_folder, proxy.order_count if duplicate else 1)
+        order_count = proxy.order_count if duplicate else 1
+        os_helpers.duplicate_file(proxy_path, destination_folder, order_count)
 
     return errors["missing"], errors["duplicates"]